+++ packages/appinio_social_share/android/src/main/java/com/appinio/socialshare/appinio_social_share/utils/SocialShareUtil.java
@@ -13,6 +13,7 @@ import android.net.Uri;
 import android.os.Build;
 import android.os.Bundle;
 import android.provider.Settings;
+import android.provider.Telephony;
 
 import androidx.core.content.FileProvider;
 
@@ -66,36 +67,52 @@ public class SocialShareUtil {
         return shareFileAndTextToPackage(imagePath, msg, context, WHATSAPP_PACKAGE);
     }
 
+    public String shareToWhatsAppFiles(ArrayList<String> imagePaths, Context context) {
+        return shareFilesToPackage(imagePaths, context, WHATSAPP_PACKAGE);
+    }
+
 
     public String shareToInstagramDirect(String text, Context activity) {
         return shareTextToPackage(text, activity, INSTAGRAM_PACKAGE);
     }
 
-    public String shareToInstagramFeed(String imagePath, Context activity, String text) {
+    public String shareToInstagramFeed(String imagePath, String message, Context activity, String text) {
         return shareFileAndTextToPackage(imagePath, text, activity, INSTAGRAM_PACKAGE);
     }
 
-    public String shareToTikTok(String imagePath, Context activity, String text) {
-        return shareFileAndTextToPackage(imagePath, text, activity, TIKTOK_PACKAGE);
+    public String shareToInstagramFeedFiles(ArrayList<String> imagePaths, Context activity, String text) {
+        return shareFilesToPackage(imagePaths, activity, INSTAGRAM_PACKAGE);
+    }
+
+    public String shareToTikTok(ArrayList<String> imagePaths, Context activity) {
+        return shareFilesToPackage(imagePaths, activity, TIKTOK_PACKAGE);
     }
 
     public String shareToTwitter(String imagePath, Context activity, String text) {
         return shareFileAndTextToPackage(imagePath, text, activity, TWITTER_PACKAGE);
     }
 
+    public String shareToTwitterFiles(ArrayList<String> imagePaths, Context activity) {
+        return shareFilesToPackage(imagePaths, activity, TWITTER_PACKAGE);
+    }
+
     public String shareToTelegram(String imagePath, Context activity, String text) {
         return shareFileAndTextToPackage(imagePath, text, activity, TELEGRAM_PACKAGE);
+    }
+
+    public String shareToTelegramFiles(ArrayList<String> imagePaths, Context activity) {
+        return shareFilesToPackage(imagePaths, activity, TELEGRAM_PACKAGE);
     }
 
 
     public String shareToMessenger(String text, Context activity) {
         Map<String, Boolean> apps = getInstalledApps(activity);
         String packageName;
-        if(apps.get("messenger")){
+        if (apps.get("messenger")) {
             packageName = FACEBOOK_MESSENGER_PACKAGE;
-        }else if(apps.get("messenger-lite")){
+        } else if (apps.get("messenger-lite")) {
             packageName = FACEBOOK_MESSENGER_LITE_PACKAGE;
-        }else{
+        } else {
             return ERROR_APP_NOT_AVAILABLE;
         }
         return shareTextToPackage(text, activity, packageName);
@@ -114,30 +131,68 @@ public class SocialShareUtil {
     }
 
     public String shareToSMS(String content, Context activity, String imagePath) {
-        String defaultApplication = Settings.Secure.getString(activity.getContentResolver(), SMS_DEFAULT_APPLICATION);
+        String defaultApplication;
+        if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.KITKAT) {
+            defaultApplication = Telephony.Sms.getDefaultSmsPackage(activity);
+        } else {
+            defaultApplication = Settings.Secure.getString(activity.getContentResolver(), SMS_DEFAULT_APPLICATION);
+        }
         return shareFileAndTextToPackage(imagePath, content, activity, defaultApplication);
+    }
+    public String shareToSMSFiles( Context activity, ArrayList<String> imagePaths) {
+        String defaultApplication;
+        if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.KITKAT) {
+            defaultApplication = Telephony.Sms.getDefaultSmsPackage(activity);
+        } else {
+            defaultApplication = Settings.Secure.getString(activity.getContentResolver(), SMS_DEFAULT_APPLICATION);
+        }
+        return shareFilesToPackage(imagePaths, activity, defaultApplication);
     }
 
 
-    public String shareToSystem(String title, String text, String filePath, String fileType, String chooserTitle, Context activity) {
+    public String shareToSystemFiles(String title, ArrayList<String> filePaths, String chooserTitle, Context activity) {
         try {
+            if (filePaths == null || filePaths.isEmpty()) return "No files to share";
+            Intent intent = new Intent();
+            intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
+            intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+            intent.setAction(Intent.ACTION_SEND_MULTIPLE);
+            ArrayList<Uri> files = new ArrayList<Uri>();
+            for (int i = 0; i < filePaths.size(); i++) {
+                Uri fileUri = FileProvider.getUriForFile(activity, activity.getApplicationContext().getPackageName() + ".provider", new File(filePaths.get(i)));
+                files.add(fileUri);
+            }
+            intent.setType(getMimeTypeOfFile(filePaths.get(0)));
+            intent.putParcelableArrayListExtra(Intent.EXTRA_STREAM, files);
+            intent.putExtra(Intent.EXTRA_SUBJECT, title);
+            intent.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION);
+            Intent chooserIntent = Intent.createChooser(intent, chooserTitle);
+            chooserIntent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
+            chooserIntent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+            activity.startActivity(chooserIntent);
+            return SUCCESS;
+        } catch (Exception e) {
+            return e.getLocalizedMessage();
+        }
+    }
 
+
+    public String shareToSystem(String title, String message, String filePath, String chooserTitle, Context activity) {
+        try {
             Intent intent = new Intent();
             intent.setFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
             intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
             intent.setAction(Intent.ACTION_SEND);
-            if (filePath != null && !filePath.isEmpty()) {
-                File file = new File(filePath);
-                Uri fileUri = FileProvider.getUriForFile(activity, activity.getApplicationContext().getPackageName() + ".provider", file);
-                intent.setType(fileType);
+            if (filePath != null) {
+                Uri fileUri = FileProvider.getUriForFile(activity, activity.getApplicationContext().getPackageName() + ".provider", new File(filePath));
+                intent.setType(getMimeTypeOfFile(filePath));
                 intent.putExtra(Intent.EXTRA_STREAM, fileUri);
             } else {
-                intent.setType("text/*");
+                intent.setType("text/plain");
             }
 
-
+            intent.putExtra(Intent.EXTRA_TEXT, message);
             intent.putExtra(Intent.EXTRA_SUBJECT, title);
-            intent.putExtra(Intent.EXTRA_TEXT, text);
             intent.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION);
             Intent chooserIntent = Intent.createChooser(intent, chooserTitle);
             chooserIntent.setFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
@@ -158,7 +213,7 @@ public class SocialShareUtil {
             shareIntent.setType("image/*");
             shareIntent.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION);
             shareIntent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-            if(stickerImage!=null){
+            if (stickerImage != null) {
                 File file = new File(stickerImage);
                 Uri stickerImageUri = FileProvider.getUriForFile(activity, activity.getApplicationContext().getPackageName() + ".provider", file);
                 shareIntent.putExtra("interactive_asset_uri", stickerImageUri);
@@ -183,8 +238,9 @@ public class SocialShareUtil {
     }
 
 
-    public void shareToFacebook(String imagePath, String text, Activity activity, MethodChannel.Result result) {
-        FacebookSdk.sdkInitialize(activity.getApplicationContext());
+    public void shareToFacebook(List<String> filePaths, String text, Activity activity, MethodChannel.Result result) {
+        FacebookSdk.fullyInitialize();
+        FacebookSdk.setApplicationId(getFacebookAppId(activity));
         callbackManager = callbackManager == null ? CallbackManager.Factory.create() : callbackManager;
         ShareDialog shareDialog = new ShareDialog(activity);
         shareDialog.registerCallback(callbackManager, new FacebookCallback<Sharer.Result>() {
@@ -205,10 +261,11 @@ public class SocialShareUtil {
                 result.success(error.getLocalizedMessage());
             }
         });
-        File file = new File(imagePath);
-        Uri fileUri = FileProvider.getUriForFile(activity, activity.getApplicationContext().getPackageName() + ".provider", file);
         List<SharePhoto> sharePhotos = new ArrayList<>();
-        sharePhotos.add(new SharePhoto.Builder().setImageUrl(fileUri).build());
+        for (int i = 0; i < filePaths.size(); i++) {
+            Uri fileUri = FileProvider.getUriForFile(activity, activity.getPackageName() + ".provider", new File(filePaths.get(i)));
+            sharePhotos.add(new SharePhoto.Builder().setImageUrl(fileUri).build());
+        }
         SharePhotoContent content = new SharePhotoContent.Builder()
                 .setShareHashtag(new ShareHashtag.Builder().setHashtag(text).build())
                 .setPhotos(sharePhotos)
@@ -222,11 +279,11 @@ public class SocialShareUtil {
         try {
             Map<String, Boolean> apps = getInstalledApps(activity);
             String packageName;
-            if(apps.get("facebook")){
+            if (apps.get("facebook")) {
                 packageName = FACEBOOK_PACKAGE;
-            }else if(apps.get("facebook-lite")){
+            } else if (apps.get("facebook-lite")) {
                 packageName = FACEBOOK_LITE_PACKAGE;
-            }else{
+            } else {
                 return ERROR_APP_NOT_AVAILABLE;
             }
 
@@ -237,7 +294,7 @@ public class SocialShareUtil {
             intent.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION);
             intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
             intent.putExtra("com.facebook.platform.extra.APPLICATION_ID", appId);
-            if(stickerImage!=null) {
+            if (stickerImage != null) {
                 File file = new File(stickerImage);
                 Uri stickerImageFile = FileProvider.getUriForFile(activity, activity.getPackageName() + ".provider", file);
                 intent.putExtra("interactive_asset_uri", stickerImageFile);
@@ -246,7 +303,7 @@ public class SocialShareUtil {
             intent.putExtra("content_url", attributionURL);
             intent.putExtra("top_background_color", backgroundTopColor);
             intent.putExtra("bottom_background_color", backgroundBottomColor);
-            if(backgroundImage!=null){
+            if (backgroundImage != null) {
                 File file1 = new File(backgroundImage);
                 Uri backgroundImageUri = FileProvider.getUriForFile(activity, activity.getPackageName() + ".provider", file1);
                 intent.setDataAndType(backgroundImageUri, getMimeTypeOfFile(backgroundImage));
@@ -283,22 +340,46 @@ public class SocialShareUtil {
     }
 
 
-    private String shareFileAndTextToPackage(String imagePath, String text, Context activity, String packageName) {
-        Intent shareIntent = new Intent(Intent.ACTION_SEND);
-
-        if (imagePath != null) {
-            File file = new File(imagePath);
-            Uri fileUri = FileProvider.getUriForFile(activity, activity.getApplicationContext().getPackageName() + ".provider", file);
-            shareIntent.putExtra(Intent.EXTRA_STREAM, fileUri);
+    private String shareFilesToPackage(List<String> imagePaths, Context activity, String packageName) {
+        if (imagePaths == null || imagePaths.isEmpty()) return "No files to share";
+        Intent shareIntent = new Intent(Intent.ACTION_SEND_MULTIPLE);
+        ArrayList<Uri> files = new ArrayList<Uri>();
+        for (int i = 0; i < imagePaths.size(); i++) {
+            Uri fileUri = FileProvider.getUriForFile(activity, activity.getApplicationContext().getPackageName() + ".provider", new File(imagePaths.get(i)));
+            files.add(fileUri);
         }
+        shareIntent.putParcelableArrayListExtra(Intent.EXTRA_STREAM, files);
+        shareIntent.setType(getMimeTypeOfFile(imagePaths.get(0)));
+        shareIntent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+        shareIntent.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION);
+        shareIntent.setPackage(packageName);
+//        if (packageName.equals(INSTAGRAM_PACKAGE) && Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+//            shareIntent.setComponent(ComponentName.createRelative(packageName, "com.instagram.share.handleractivity.ShareHandlerActivity")); //open instagram feed
+//        }
+        try {
+            activity.startActivity(shareIntent);
+            return SUCCESS;
+        } catch (Exception e) {
+            e.printStackTrace();
+            return e.getLocalizedMessage();
+        }
+    }
 
-        shareIntent.setType(imagePath == null ? "text/*" : getMimeTypeOfFile(imagePath));
-        shareIntent.putExtra(Intent.EXTRA_TEXT, text);
+    private String shareFileAndTextToPackage(String imagePath, String message, Context activity, String packageName) {
+        Intent shareIntent = new Intent(Intent.ACTION_SEND);
+        if (imagePath != null) {
+            Uri fileUri = FileProvider.getUriForFile(activity, activity.getApplicationContext().getPackageName() + ".provider", new File(imagePath));
+            shareIntent.putExtra(Intent.EXTRA_STREAM, fileUri);
+            shareIntent.setType(getMimeTypeOfFile(imagePath));
+        } else {
+            shareIntent.setType("text/plain");
+        }
+        shareIntent.putExtra(Intent.EXTRA_TEXT, message);
         shareIntent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
         shareIntent.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION);
         shareIntent.setPackage(packageName);
         if (packageName.equals(INSTAGRAM_PACKAGE) && Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-            shareIntent.setComponent(ComponentName.createRelative(packageName,"com.instagram.share.handleractivity.ShareHandlerActivity")); //open instagram feed
+            shareIntent.setComponent(ComponentName.createRelative(packageName, "com.instagram.share.handleractivity.ShareHandlerActivity")); //open instagram feed
         }
         try {
             activity.startActivity(shareIntent);
@@ -332,7 +413,7 @@ public class SocialShareUtil {
         intent.setData(Uri.parse("sms:"));
         List<ResolveInfo> resolvedActivities = pm.queryIntentActivities(intent, 0);
         apps.put("message", !resolvedActivities.isEmpty());
-        String[] appNames = {"instagram", "facebook_stories", "whatsapp", "telegram", "messenger", "facebook","facebook-lite","messenger-lite", "instagram_stories", "twitter", "tiktok"};
+        String[] appNames = {"instagram", "facebook_stories", "whatsapp", "telegram", "messenger", "facebook", "facebook-lite", "messenger-lite", "instagram_stories", "twitter", "tiktok"};
 
         for (int i = 0; i < appNames.length; i++) {
             try {
@@ -346,7 +427,7 @@ public class SocialShareUtil {
     }
 
     private static String getMimeTypeOfFile(String pathName) {
-        try{
+        try {
             Path path;
             if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.O) {
                 path = new File(pathName).toPath();
@@ -354,13 +435,13 @@ public class SocialShareUtil {
                 return mimeType;
             }
             return "*/*";
-        }catch (Exception e){
+        } catch (Exception e) {
             return "";
         }
     }
 
 
-    String getFacebookAppId(Context activity){
+    String getFacebookAppId(Context activity) {
         String appId = "";
         try {
             ApplicationInfo appInfo = activity.getPackageManager().getApplicationInfo(
@@ -369,7 +450,7 @@ public class SocialShareUtil {
             Bundle metaData = appInfo.metaData;
             if (metaData != null) {
                 appId = metaData.getString("com.facebook.sdk.ApplicationId");
-                Log.d("FB_APP_ID",appId);
+                Log.d("FB_APP_ID", appId);
             }
         } catch (PackageManager.NameNotFoundException e) {
             // Handle the exception if needed