+++ packages/appinio_social_share/example/lib/main.dart
@@ -1,3 +1,4 @@
+import 'package:file_picker/file_picker.dart';
 import 'package:flutter/material.dart';
 import 'package:appinio_social_share/appinio_social_share.dart';
 
@@ -21,10 +22,17 @@ class _MyAppState extends State<MyApp> {
         title: "Share Feature",
         home: Scaffold(
           body: Column(
+            mainAxisAlignment: MainAxisAlignment.center,
             children: [
               ElevatedButton(
                 child: const Text("ShareToWhatsapp"),
-                onPressed: () {},
+                onPressed: () async {
+                  FilePickerResult? result = await FilePicker.platform
+                      .pickFiles(type: FileType.image, allowMultiple: false);
+                  if (result != null && result.paths.isNotEmpty) {
+                    shareToWhatsApp("message", result.paths[0]!);
+                  }
+                },
               ),
             ],
           ),
@@ -32,6 +40,6 @@ class _MyAppState extends State<MyApp> {
   }
 
   shareToWhatsApp(String message, String filePath) async {
-    await appinioSocialShare.shareToWhatsapp(message, filePath: filePath);
+    await appinioSocialShare.android.shareToSMS(message, filePath);
   }
 }