+++ packages/appinio_social_share/lib/appinio_social_share_platform_interface.dart
@@ -28,12 +28,12 @@ abstract class AppinioSocialSharePlatform extends PlatformInterface {
     throw UnimplementedError('platformVersion() has not been implemented.');
   }
 
-  Future<Map> getInstalledApps() {
+  Future<Map<String, bool>> getInstalledApps() {
     throw UnimplementedError('platformVersion() has not been implemented.');
   }
 
   ///filaPath only works for android
-  Future<String> shareToWhatsapp(String message, {String? filePath}) {
+  Future<String> shareToWhatsapp(String message, {List<String>? filePaths}) {
     throw UnimplementedError('platformVersion() has not been implemented.');
   }
 
@@ -41,7 +41,11 @@ abstract class AppinioSocialSharePlatform extends PlatformInterface {
     throw UnimplementedError('platformVersion() has not been implemented.');
   }
 
-  Future<String> shareToInstagramFeed(String filePath) {
+  Future<String> shareToInstagramFeed(String message, String? filePath) async {
+    throw UnimplementedError('platformVersion() has not been implemented.');
+  }
+
+  Future<String> shareToInstagramFeedAndroid(List<String> filePaths) async {
     throw UnimplementedError('platformVersion() has not been implemented.');
   }
 
@@ -50,25 +54,26 @@ abstract class AppinioSocialSharePlatform extends PlatformInterface {
   }
 
   ///This function only works for android
-  Future<String> shareToTiktokStatus(String filePath) {
+  Future<String> shareToTiktokStatus(List<String> filePaths) {
     throw UnimplementedError('platformVersion() has not been implemented.');
   }
 
   ///This function only works for iOS
-  Future<String> shareToTiktokPost(String videoFile) {
+  Future<String> shareToTiktokPost(
+      String filePath, String redirectUrl, TiktokFileType tiktokFileType) {
     throw UnimplementedError('platformVersion() has not been implemented.');
   }
 
-  Future<String> shareToTwitter(String message, {String? filePath}) {
+  Future<String> shareToTwitter(String message, String? filePath) {
     throw UnimplementedError('platformVersion() has not been implemented.');
   }
 
   ///filepath only works for android
-  Future<String> shareToTelegram(String message, {String? filePath}) {
+  Future<String> shareToTelegram(String message, {List<String>? filePaths}) {
     throw UnimplementedError('platformVersion() has not been implemented.');
   }
 
-  Future<String> shareToFacebook(String hashtag, String filePath) {
+  Future<String> shareToFacebook(String hashtag, List<String> filePaths) {
     throw UnimplementedError('platformVersion() has not been implemented.');
   }
 
@@ -97,12 +102,71 @@ abstract class AppinioSocialSharePlatform extends PlatformInterface {
   }
 
   Future<String> shareToSystem(String title, String message,
-      {String? filePath}) {
+      {List<String>? filePaths}) {
     throw UnimplementedError('platformVersion() has not been implemented.');
   }
 
   ///filePath only works for android
-  Future<String> shareToSMS(String message, {String? filePath}) {
+  Future<String> shareToSMS(String message, {List<String>? filePaths}) {
     throw UnimplementedError('platformVersion() has not been implemented.');
   }
+
+  Future<String> shareToWhatsappAndroid(
+      String message, String? filePath) async {
+    throw UnimplementedError('platformVersion() has not been implemented.');
+  }
+
+  Future<String> shareToWhatsappAndroidMultiFiles(
+      List<String> filePaths) async {
+    throw UnimplementedError('platformVersion() has not been implemented.');
+  }
+
+  Future<String> shareToTelegramAndroid(
+      String message, String? filePath) async {
+    throw UnimplementedError('platformVersion() has not been implemented.');
+  }
+
+  Future<String> shareToTelegramAndroidMultiFiles(
+      List<String> filePaths) async {
+    throw UnimplementedError('platformVersion() has not been implemented.');
+  }
+
+  Future<String> shareToTwitterAndroid(String message, String? filePath) async {
+    throw UnimplementedError('platformVersion() has not been implemented.');
+  }
+
+  Future<String> shareToTwitterAndroidMultifiles(List<String> filePaths) async {
+    throw UnimplementedError('platformVersion() has not been implemented.');
+  }
+
+  Future<String> shareToSMSAndroid(String message, String? filePath) async {
+    throw UnimplementedError('platformVersion() has not been implemented.');
+  }
+
+  Future<String> shareToSMSAndroidMultifiles(List<String> filePaths) async {
+    throw UnimplementedError('platformVersion() has not been implemented.');
+  }
+
+  Future<String> shareToSystemAndroid(
+      String title, String message, String? filePath) async {
+    throw UnimplementedError('platformVersion() has not been implemented.');
+  }
+
+  Future<String> shareImageToWhatsApp(String filePath) async {
+    throw UnimplementedError('platformVersion() has not been implemented.');
+  }
+
+  Future<String> shareToSystemAndroidMultifiles(
+      String title, List<String> filePaths) async {
+    throw UnimplementedError('platformVersion() has not been implemented.');
+  }
+}
+
+enum TiktokFileType {
+  image("image"),
+  video("video");
+
+  final String value;
+
+  const TiktokFileType(this.value);
 }