+++ packages/appinio_social_share/CHANGELOG.md
@@ -1,3 +1,19 @@
+## [0.3.2] - 2024.08.01
+* Many bug fixes and improvements.
+* Version upgrades for dependencies.
+
+## [0.3.1] - 2024.04.15
+* Many bug fixes and improvements.
+* Refactoring:
+  - Now you will have to choose platform while sharing.
+  - For example: AppinioSocialShare().android.shareToWhatsapp(message, file)
+
+## [0.3.0] - 2023.12.21
+* Enhancement:
+  - Facebook Android Sdk updated to 16.0.0.  
+* Features:
+  - Now you can share multiple files to whatsapp, telegram, facebook and other apps.
+
 ## [0.2.6] - 2023.09.25
 
 * Enhancements: